moving home screen into seperate component
